+++ erde/tokenize.lua
@@ -53,7 +53,7 @@ local function Newline()
   return consume()
 end
 
-local function EscapeChar(preventInterpolation)
+local function EscapeChar(preventInterpolation, preventInvalidEscape)
   consume() -- backslash
 
   if not preventInterpolation and (char == '{' or char == '}') then
@@ -105,6 +105,8 @@ local function EscapeChar(preventInterpolation)
     end
 
     return escapeChar .. consume()
+  elseif preventInvalidEscape then
+    return '\\' .. consume()
   else
     throw('invalid escape sequence \\' .. char)
   end
@@ -301,7 +303,7 @@ local function LongString()
     elseif char == '\n' then
       content = content .. Newline()
     elseif char == '\\' then
-      content = content .. EscapeChar()
+      content = content .. EscapeChar(false, true)
     elseif char == '{' then
       if content ~= '' then
         commit(content)